requirements.txt
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.10.0
beautifulsoup4==4.14.2
certifi==2025.10.5
charset-normalizer==3.4.3
click==8.3.0
exceptiongroup==1.3.0
fastapi==0.116.2
googleapis-common-protos==1.70.0
h11==0.16.0
idna==3.10
importlib_metadata==8.7.0
iniconfig==2.1.0
joblib==1.5.2
numpy==1.26.4
opentelemetry-api==1.37.0
opentelemetry-distro==0.58b0
opentelemetry-exporter-otlp-proto-common==1.37.0
opentelemetry-exporter-otlp-proto-http==1.37.0
opentelemetry-instrumentation==0.58b0
opentelemetry-instrumentation-asgi==0.58b0
opentelemetry-instrumentation-fastapi==0.58b0
opentelemetry-instrumentation-requests==0.58b0
opentelemetry-proto==1.37.0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
opentelemetry-util-http==0.58b0
packaging==25.0
pluggy==1.6.0
polars==1.34.0
polars-runtime-32==1.34.0
protobuf==6.32.1
psycopg2-binary==2.9.10
pydantic==2.12.0
pydantic_core==2.41.1
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.2
python-dotenv==1.1.1
requests==2.32.5
scikit-learn==1.5.2
scipy==1.15.3
sniffio==1.3.1
soupsieve==2.8
starlette==0.48.0
threadpoolctl==3.6.0
tomli==2.2.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.35.0
wrapt==1.17.3
zipp==3.23.0
